requirements.txt
absl-py==2.1.0
accelerate==0.31.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
Brotli @ file:///croot/brotli-split_1714483155106/work
build==1.2.2.post1
CacheControl==0.14.1
certifi @ file:///croot/certifi_1725551672989/work/certifi
cffi==1.17.1
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
cleo==2.1.0
cmake==3.30.5
contourpy==1.3.0
crashtest==0.4.1
cryptography==43.0.3
cycler==0.12.1
distlib==0.3.9
dulwich==0.21.7
fastjsonschema==2.20.0
filelock @ file:///croot/filelock_1700591183607/work
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.9.0
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645438755360/work
grpcio==1.66.1
h5py==3.11.0
huggingface-hub==0.25.1
idna @ file:///croot/idna_1714398848350/work
importlib_metadata==8.5.0
importlib_resources==6.4.5
installer==0.7.0
jaraco.classes==3.4.0
jeepney==0.8.0
Jinja2 @ file:///croot/jinja2_1716993405101/work
joblib==1.4.2
keyring==24.3.1
kiwisolver==1.4.7
kogger==0.1.2
libigl==2.5.1
lit==18.1.8
Markdown==3.7
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
matplotlib==3.9.2
mkl-service==2.4.0
mkl_fft @ file:///croot/mkl_fft_1725370245198/work
mkl_random @ file:///croot/mkl_random_1725370241878/work
more-itertools==10.5.0
mpmath @ file:///croot/mpmath_1690848262763/work
msgpack==1.1.0
multidict==6.1.0
networkx @ file:///croot/networkx_1717597493534/work
numpy==1.26.4
packaging==24.1
pexpect==4.9.0
pillow @ file:///croot/pillow_1721059439630/work
pkginfo==1.11.2
platformdirs==4.3.6
poetry==1.8.4
poetry-core==1.9.1
poetry-plugin-export==1.8.0
protobuf==5.28.2
psutil==6.0.0
ptyprocess==0.7.0
pycparser==2.22
pyg-lib==0.4.0+pt24cu118
pyparsing==3.1.4
pyproject_hooks==1.2.0
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil==2.9.0.post0
PyYAML @ file:///croot/pyyaml_1698096049011/work
RapidFuzz==3.10.1
requests @ file:///croot/requests_1721410876868/work
requests-toolbelt==1.0.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
SecretStorage==3.3.3
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
sympy==1.12
tensorboard==2.17.1
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tomli==2.1.0
tomlkit==0.13.2
torch==2.4.1
torch-geometric==2.6.0
torch_cluster==1.6.3+pt24cu118
torch_scatter==2.1.2+pt24cu118
torch_sparse==0.6.18+pt24cu118
torch_spline_conv==1.2.2+pt24cu118
torchaudio==2.4.1
torchvision==0.19.1
tqdm==4.66.5
triton==2.0.0
trove-classifiers==2024.10.21.16
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
urllib3 @ file:///croot/urllib3_1718912636303/work
virtualenv==20.28.0
Werkzeug==3.0.4
yarl==1.11.1
zipp==3.20.2
